requirements.txt
asgiref==3.2.3
boto3==1.11.4
botocore==1.14.4
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.2
django-heroku==0.3.1
django-storages==1.8
docutils==0.15.2
dropbox==9.4.0
gunicorn==20.0.4
idna==2.8
Jinja2==2.10.3
jmespath==0.9.4
MarkupSafe==1.1.1
psycopg2==2.8.4
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2019.3
requests==2.22.0
s3transfer==0.3.1
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.7
whitenoise==5.0.1
